R4R: Update cosmos-sdk version (#12)
* update sdk version

* fix simulation test

* fix typo

* define MsgTransferTokenOwner used by protobuf

* add yaml tag

* increase code start at 2

+++ exported/exported.go
@@ -7,7 +7,7 @@ import (
 type TokenI interface {
 	GetSymbol() string
 	GetName() string
-	GetScale() uint8
+	GetScale() uint32
 	GetMinUnit() string
 	GetInitialSupply() uint64
 	GetMaxSupply() uint64
